Skills created, fireball is available q+up|right

+++ core/src/com/gdx/game/factories/Weapon.java
@@ -14,10 +14,10 @@ public class Weapon {
 
     private final Bullet bullet;
     private final Entity shooter;
-    private final int shootingRate;
+    private final float shootingRate;
     private Date lastFireDate = null;
 
-    public Weapon(Entity shooterEntity, Bullet bullet, int shootingRate) {
+    public Weapon(Entity shooterEntity, Bullet bullet, float shootingRate) {
         this.shooter = shooterEntity;
         this.bullet = bullet;
         this.shootingRate = shootingRate;
